LayoutTests/imported/w3c/web-platform-tests/css/css-transforms/transform-scale-hittest-expected.txt


PASS Hit test within unscaled box
FAIL Hit test intersecting scaled box assert_equals: expected Element node <div id="scaled"></div> but got Element node <div id="normal"></div>
